Fixed bug with new README file.
Dist::Zilla was adding a README file itself, so it complained that one
already existed.

Now skipping the existing README file (which I added to make Github
happy) during 'dzil build' so that the README file here is for Github
but the README file in the distribution will be generated by Dist::Zilla
from the dist.ini file.