Fixed bug in Traversal::dryrun()
dryrun() requires that the visit attribute be read/write so that it can
temporarily use set_visit() to set the visit attribute to the
appropriately "dry" version.

I also took the opportunity to change the attribute to be non-required.
Later on in the test for Traversal I will use this, as I will only use
the dryrun() mode, so it makes sense to not have to set a visit CodeRef
at all.