Added 03-traversal.t, which tests Traversal::run()
This test is heavily data driven and has a mini data-structure-y DSL to
help to set up test cases to run. It utilises the dryrun() method of
Traversal, which is a wrapper around the run() method so it does test
the primary behaviour.

Also updated in this commit is dist.ini, because two additional
requirements were added for the new test:

Test::Deep is used for cmp_deeply() and any(), which is useful when
there is a random element to the test output. This is the case for
example when there are multiple valid solutions and one is picked at
random. (Actually it's the hash keys that do it. I've tested this test
on Perl 5.17 to ensure that properly random hash keys work as expected.)

Test::Exception is used to make sure that a cyclical graph causes an
exception to be thrown if we attempt to traverse it.