Added test for redundant edge removal.
This test is in traversal.t although it is really testing the behaviour
of the private method Algorithm::DependencySolver::Solver::_remove_redundancy()

For this test I added a test option 'extra_tests' which can be a
subroutine reference that gets called and in which it is expected that
additional ok() calls will be made. This allows us to check facts about
the result *other* than the resultant traversal.